OpenJDK integration.
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@4796 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/openjdk/java/java/util/concurrent/atomic/AtomicLong.java
@@ -58,7 +58,7 @@ public class AtomicLong extends Number implements java.io.Serializable {
 
     /**
      * Records whether the underlying JVM supports lockless
-     * CompareAndSet for longs. While the unsafe.CompareAndSetLong
+     * compareAndSwap for longs. While the Unsafe.compareAndSwapLong
      * method works in either case, some constructions should be
      * handled at Java level to avoid locking user-visible locks.
      */
@@ -265,7 +265,7 @@ public class AtomicLong extends Number implements java.io.Serializable {
     }
 
     public long longValue() {
-	return (long)get();
+        return get();
     }
 
     public float floatValue() {
